1.各星球分颜色显示血条 2.优化游戏加载流程 3.LineRenderManager 4.优化部分UI

+++ Assets/ResourceUnit.cs
@@ -13,7 +13,7 @@ public enum ResourceType
 public struct ResourceTable
 {
     public ResourceType resourceType;
-    public float resourceNum;
+    public int resourceNum;
 }
 public class ResourceUnit : MonoBehaviour
 {
